+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added demangle tool.
 - Added uncc tool.
 - The end of a function in the symbol table is now identified correctly so that extra local variables are not erroneously generated.
+- The type deduplication algorithm has been improved.
 - stdump: Fixed a bug where global variables would be printed to stdout even if an output file was specified.
 
 ## v1.0